also export the entire line
It might be useful for users of this fact to have access to the entire
original public key, comments, warts and all.

For example, some people (like me) might be unhappy about the built-in
ssh_authorized_key type and might have decided to implement their own
type, based on a simple concat::fragment. In that case, it's much
simpler to embed simply the entire public key than reassemble it by
type/key/comment.

In fact, it's unclear to me why the original ssh_authorized_key
resource split it up like this, in retrospect. But I understand it's
necessary for this fact to export that to have it backwards
compatible, so that's left.