replace deprecated Modulefile with metadata.json
Modulefile was deprecated during the puppet 3.x cycle.

The file also mentions that the license was Apache 2.0 but the README
for this repos says the code is GPL3. I've reused GPL3 since this is
what I was assuming this code to be licensed under.

I haven't explicitly added FreeBSD, OpenBSD and Gentoo support since I
think no one participating in this module has been using the module for
these distros/platforms for a good while. We might want to drop support
for them, or add them explicitly if someone manifests interest in keeping
them.